requirements.txt
click>=8.1.0
yt-dlp>=2024.0.0
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
biliup>=0.3.0
openai>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
pillow>=10.0.0
